+++ src/shelly_analyzer/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "13.10.1"
+__version__ = "14.0.0"